created individual components for each section
