+++ src/components/HomepageHeader/styles.module.scss
@@ -1,0 +1,63 @@
+/*** Hero banner section ***/
+.heroBanner {
+    min-height: 90vh;
+    min-height: 90svh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding: 2rem 0;
+
+    &__title {
+        max-width: 25ch;
+        font-size: var(--fs-hero);
+        line-height: var(--lh-hero);
+        text-align: center;
+        margin-bottom: 2rem;
+        transform: translateY(1rem);
+        opacity: 0;
+        animation: fadeIn 1s ease forwards;
+    }
+
+    &__subheader {
+        text-align: center;
+        transform: translateY(1rem);
+        opacity: 0;
+        animation: fadeIn 1s ease forwards 250ms;
+    }
+}
+
+.bannerCta {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 1rem;
+    width: 100%;
+    margin-top: 4rem;
+    transform: translateY(1rem);
+    opacity: 0;
+    animation: fadeIn 1s ease forwards 500ms;
+
+    a {
+        font-size: var(--fs-body);
+    }
+}
+
+@media screen and (min-width: 500px) {
+
+    .bannerCta {
+        flex-direction: row;
+
+        a {
+            width: fit-content;
+        }
+    }
+}
+
+@keyframes fadeIn {
+    to {
+        transform: translateY(0);
+        opacity: 1;
+    }
+}
