+++ src/components/ServicesSection/styles.module.scss
@@ -2,14 +2,6 @@
 .servicesSection {
   padding: 4rem 0;
 
-  &__container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    gap: 2rem;
-  }
-
   &__title,
   &__subheader {
     text-align: center;
@@ -32,27 +24,29 @@
     flex-direction: column;
     align-items: flex-start;
     justify-content: space-between;
-    background-color: var(--neutral-clr-900);
+    background-color: #fffffe;
+    border: 1px solid #e4e6e6;
+    box-shadow: 1px 1px 10px rgba(0, 0, 0, 0.1);
 
     &__icon {
       width: 100%;
       text-align: right;
+      color: #fffffe;
 
       .icon {
         padding: .75rem;
         border-radius: 10px;
         background-color: var(--ifm-color-primary);
-        border: 1px solid var(--neutral-clr-50);
-        box-shadow: 2px 2px 0 var(--neutral-clr-50);
+        border: 1px solid #cacccc;
+        box-shadow: 2px 2px 0 #cacccc;
         font-size: var(--fs-icon);
       }
     }
 
-    &__icon,
     &__title,
     &__description,
     .tag {
-      color: var(--neutral-clr-50);
+      color: #141414;
     }
 
     &__title {
@@ -74,7 +68,7 @@
 
       .tag {
         padding: .25rem .75rem;
-        background-color: var(--neutral-clr-700);
+        background-color: #cacccc;
         border-radius: 5px;
         font-size: var(--fs-footer);
         font-weight: var(--fw-medium);
@@ -97,6 +91,7 @@
 
 @media screen and (min-width: 768px) {
   .servicesSection {
+
     .thing,
     .protocol {
       width: 25rem;
