+++ src/css/custom.scss
@@ -143,7 +143,6 @@ html {
   background-color: var(--neutral-clr-50);
 }
 
-
 body,
 p,
 button,
@@ -153,28 +152,6 @@ a {
   letter-spacing: var(--ls-body);
   color: var(--neutral-clr-900);
   margin: 0;
-}
-
-.gridBackground {
-  position: absolute;
-  top: 60px;
-  left: 0;
-  width: 100%;
-  z-index: -1;
-  mask: linear-gradient(135deg, rgba(255, 255, 255, 0.5) 5%,rgba(255, 255, 255, .25) 20%,rgba(255, 255, 255, .05) 35%, rgba(255, 255, 255, 0) 50%, transparent);
-  overflow: hidden;
-
-  svg {
-    max-width: 100rem;
-    max-height: 100rem;
-    min-width: 65rem;
-    min-height: 65rem;
-  }
-}
-
-.hero-title {
-  font-family: var(--ff-hero);
-  letter-spacing: var(--ls-hero);
 }
 
 h1,
@@ -202,6 +179,178 @@ h3 {
   letter-spacing: var(--ls-header)
 }
 
+a,
+a:hover {
+  text-decoration: none;
+}
+
+ul {
+  list-style: none;
+}
+
+.icon {
+  font-size: var(--fs-body-lg);
+  // filter: drop-shadow(2px 2px var(--neutral-clr-900));
+}
+
+/*
+img {
+  width: 100%;
+  height: 100%;
+} */
+
+/************************************/
+/*** Overriding docusaurus styles ***/
+/************************************/
+
+.navbar__title {
+  font-family: var(--ff-primary);
+  font-size: var(--fs-body-sm);
+  font-weight: var(--fw-regular);
+  line-height: normal;
+}
+
+a.navbar__item {
+  font-family: var(--ff-primary);
+  font-weight: var(--fw-regular);
+}
+
+.navbar-sidebar__back {
+  font-weight: var(--fw-regular);
+}
+
+.breadcrumbs__item {
+  .breadcrumbs__link {
+    line-height: unset;
+  }
+}
+
+.menu__list {
+  .menu__link {
+    font-size: var(--fs-body-sm);
+  }
+}
+
+footer.footer {
+
+  div,
+  ul,
+  li,
+  a,
+  p {
+    font-size: var(--fs-footer);
+    line-height: 2.5;
+  }
+}
+
+.card h2 {
+  letter-spacing: var(--ls-subheader);
+}
+
+.header-github-link {
+  &::before {
+    content: '';
+    mask-image: url('/img/github.svg');
+    mask-repeat: no-repeat;
+    mask-size: cover;
+    background-color: var(--ifm-navbar-link-color);
+    background-size: cover;
+    background-repeat: no-repeat;
+    display: flex;
+    width: 1.5rem;
+    height: 1.5rem;
+    transition: background-color var(--ifm-transition-fast) var(--ifm-transition-timing-default);
+  }
+
+  &:hover::before {
+    background-color: var(--ifm-color-primary);
+  }
+
+}
+
+/********************/
+/*** Footer logos ***/
+/********************/
+
+.logos-wrapper {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: flex-start;
+  flex-wrap: wrap;
+  gap: 1rem;
+
+  .foundation-logo {
+    width: 7rem;
+  }
+
+  .iot-logo,
+  .w3c-logo {
+    width: 3.5rem;
+  }
+
+  .footer-logo:not(.w3c-logo) {
+    svg {
+      path {
+        fill: var(--neutral-clr-900);
+      }
+    }
+  }
+}
+
+.footer__copyright {
+  display: flex;
+  flex-direction: column;
+  align-items: start;
+  justify-content: space-between;
+  border-top: 1px solid var(--neutral-clr-900);
+  padding-top: 1.5rem;
+  gap: 1rem;
+
+  .thingweb-footer-logo {
+    width: 8rem;
+
+    path {
+      fill: var(--neutral-clr-900);
+    }
+  }
+
+  p {
+    text-align: left;
+  }
+}
+
+@media screen and (min-width: 997px) {
+  .logos-wrapper {
+    width: 8rem;
+    justify-content: space-between;
+
+    .foundation-logo {
+      width: 100%;
+    }
+
+    .iot-logo,
+    .w3c-logo {
+      width: calc(50% - .5rem);
+    }
+  }
+
+  .footer__copyright {
+    flex-direction: row;
+    align-items: center;
+  }
+}
+
+/***********************/
+/****** Utilities ******/
+/***********************/
+
+.hero-title {
+  font-family: var(--ff-hero);
+  letter-spacing: var(--ls-hero);
+}
+
 .hero-subheader,
 .subheader {
   line-height: var(--lh-subheader);
@@ -225,29 +374,17 @@ h3 {
   margin: 0 auto;
 }
 
-a,
-a:hover {
-  text-decoration: none;
+.text-highlight {
+  color: var(--ifm-color-primary);
 }
 
-ul {
-  list-style: none;
+.section-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 2rem;
 }
-
-.icon {
-  font-size: var(--fs-body-lg);
-  // filter: drop-shadow(2px 2px var(--neutral-clr-900));
-}
-
-/*
-img {
-  width: 100%;
-  height: 100%;
-} */
-
-/***********************/
-/****** Utilities ******/
-/***********************/
 
 
 /*** Buttons styles **/
@@ -331,35 +468,37 @@ img {
 }
 
 .btn-tertiary {
-  background-color: var(--neutral-clr-700);
-  border-color: var(--neutral-clr-700);
+  background-color: #e4e6e6;
+  border-color: #e4e6e6;
+  color: #141414;
 
   &.btn-outline {
     background-color: transparent;
-    color: var(--neutral-clr-900);
+    color: #fffffe;
   }
 
   &:hover {
-    background-color: var(--neutral-clr-600);
-    border-color: var(--neutral-clr-600);
-    color: var(--neutral-clr-50);
+    background-color: #cacccc;
+    border-color: #cacccc;
+    color: #141414;
   }
 
   &:active {
-    background-color: var(--neutral-clr-500);
-    border-color: var(--neutral-clr-500);
-    color: var(--neutral-clr-50);
+    background-color: #b1b3b3;
+    border-color: #b1b3b3;
+    color: #141414;
   }
 }
 
 /*** Scrollers ***/
+
 .scroller {
   width: 100%;
   padding: 2rem 0;
   overflow: hidden;
   mask: linear-gradient(90deg, transparent, rgb(255, 255, 255) 10%, white 90%, transparent);
 
-  & > * {
+  &>* {
     width: max-content;
     display: flex;
     flex-direction: row;
@@ -394,146 +533,13 @@ img {
   }
 }
 
+
+/********************/
+/**** Animations ****/
+/********************/
+
 @keyframes scroll {
   to {
     transform: translate(calc(-50% - (var(--_content-gap)/2)));
-  }
-}
-
-
-/*** Overriding docusaurus styles ***/
-.navbar__title {
-  font-family: var(--ff-primary);
-  font-size: var(--fs-body-sm);
-  font-weight: var(--fw-regular);
-  line-height: normal;
-}
-
-a.navbar__item {
-  font-family: var(--ff-primary);
-  font-weight: var(--fw-regular);
-}
-
-.navbar-sidebar__back {
-  font-weight: var(--fw-regular);
-}
-
-.breadcrumbs__item {
-  .breadcrumbs__link {
-    line-height: unset;
-  }
-}
-
-.menu__list {
-  .menu__link {
-    font-size: var(--fs-body-sm);
-  }
-}
-
-footer.footer {
-
-  div,
-  ul,
-  li,
-  a,
-  p {
-    font-size: var(--fs-footer);
-    line-height: 2.5;
-  }
-}
-
-.card h2 {
-  letter-spacing: var(--ls-subheader);
-}
-
-.header-github-link {
-  &::before {
-    content: '';
-    mask-image: url('/img/github.svg');
-    mask-repeat: no-repeat;
-    mask-size: cover;
-    background-color: var(--ifm-navbar-link-color);
-    background-size: cover;
-    background-repeat: no-repeat;
-    display: flex;
-    width: 1.5rem;
-    height: 1.5rem;
-    transition: background-color var(--ifm-transition-fast) var(--ifm-transition-timing-default);
-  }
-
-  &:hover::before {
-    background-color: var(--ifm-color-primary);
-  }
-
-}
-
-/*** Footer logos ***/
-.logos-wrapper {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: flex-start;
-  flex-wrap: wrap;
-  gap: 1rem;
-
-  .foundation-logo {
-    width: 7rem;
-  }
-
-  .iot-logo,
-  .w3c-logo {
-    width: 3.5rem;
-  }
-
-  .footer-logo:not(.w3c-logo) {
-    svg {
-      path {
-        fill: var(--neutral-clr-900);
-      }
-    }
-  }
-}
-
-.footer__copyright {
-  display: flex;
-  flex-direction: column;
-  align-items: start;
-  justify-content: space-between;
-  border-top: 1px solid var(--neutral-clr-900);
-  padding-top: 1.5rem;
-  gap: 1rem;
-
-  .thingweb-footer-logo {
-    width: 8rem;
-
-    path {
-      fill: var(--neutral-clr-900);
-    }
-  }
-
-  p {
-    text-align: left;
-  }
-}
-
-@media screen and (min-width: 997px) {
-  .logos-wrapper {
-    width: 8rem;
-    justify-content: space-between;
-
-    .foundation-logo {
-      width: 100%;
-    }
-
-    .iot-logo,
-    .w3c-logo {
-      width: calc(50% - .5rem);
-    }
-  }
-
-  .footer__copyright {
-    flex-direction: row;
-    align-items: center;
   }
 }
